correct minors for build

+++ pkg/controller/utils/grafana_service.go
@@ -33,7 +33,7 @@ func getServiceType(cr *v1alpha1.Grafana) corev1.ServiceType {
 }
 
 func getServicePorts(cr *v1alpha1.Grafana, currentState *corev1.Service) []corev1.ServicePort {
-	intPort := DefaultGrafanaPort
+	intPort := int32(DefaultGrafanaPort)
 
 	defaultPorts := []corev1.ServicePort{
 		{
